Workflow change
Removed previous workflows and replaced with new ones. The new workflow will be to push new component versions as a tag to master in the format <component>_<version>. This will trigger an update of the component Helm chart version and appVersion.
Then we trigger a secondary workflow that does the actual release.